xf86-video-vmware: fix build issue
Fixes this build issue:

| rmdir: failed to remove
`/srv/home/nitin/builds/build-gcc47/tmp/work/i586-poky-linux/xf86-video-vmware-2_12.0.1-r17.1/image/usr/bin':
No such file or directory
NOTE: package xf86-video-vmware-2_12.0.1-r17.1: task do_install: Failed

No PR bump as no change in the output.

+++ meta/recipes-graphics/xorg-driver/xf86-video-vmware_12.0.2.bb
@@ -18,5 +18,7 @@ COMPATIBLE_HOST = '(i.86.*-linux|x86_64.*-linux)'
 
 do_install_append () {
 	# Remove useless empty directory
-	rmdir ${D}${bindir}
+	if [ -d ${D}${bindir} ]; then
+		rmdir ${D}${bindir}
+	fi
 }
